+++ src/pages/dashboard/index.tsx
@@ -2,13 +2,13 @@ import { useState, useEffect, useMemo } from 'react';
 import Head from 'next/head';
 import { Heading, HeadingLevels } from 'lbh-frontend-react';
 import { ReactNode } from 'react';
-import { EvidenceRequest } from '../domain/evidence-request';
-import { InternalApiGateway } from '../gateways/internal-api';
-import { EvidenceRequestTable } from '../components/EvidenceRequestTable';
-import ResidentSearchForm from '../components/ResidentSearchForm';
-import Tabs from '../components/Tabs';
-import TableSkeleton from '../components/TableSkeleton';
-import Layout from '../components/DashboardLayout';
+import { EvidenceRequest } from '../../domain/evidence-request';
+import { InternalApiGateway } from '../../gateways/internal-api';
+import { EvidenceRequestTable } from '../../components/EvidenceRequestTable';
+import ResidentSearchForm from '../../components/ResidentSearchForm';
+import Tabs from '../../components/Tabs';
+import TableSkeleton from '../../components/TableSkeleton';
+import Layout from '../../components/DashboardLayout';
 
 const BrowseResidents = (): ReactNode => {
   const [evidenceRequests, setEvidenceRequests] = useState<EvidenceRequest[]>();
